Implement recording using double batteryPercent

+++ 3004F22/Recording.cpp
@@ -1,6 +1,6 @@
 #include "Recording.h"
 
-Recording::Recording(DeviceMediator* device, const string& name, int intensity, int group, int batteryPercent, int connection):
+Recording::Recording(DeviceMediator* device, const string& name, int intensity, int group, double batteryPercent, int connection):
     name(name), intensity(intensity), group(group), batteryPercent(batteryPercent), connection(connection), Widget(device) {}
 
 Recording::~Recording() { delete this; }
@@ -11,6 +11,6 @@ int Recording::getIntensity() { return this->intensity; }
 
 int Recording::getGroup() { return this->group; }
 
-int Recording::getBatteryPercent() { return this->batteryPercent; }
+double Recording::getBatteryPercent() { return this->batteryPercent; }
 
 int Recording::getConnection() { return this->connection; }
